# backend/src/main/java/com/lqq/supportflow/model/domain/ModelConfig.java
package com.lqq.supportflow.model.domain;

public record ModelConfig(
        @com.fasterxml.jackson.databind.annotation.JsonSerialize(using = com.fasterxml.jackson.databind.ser.std.ToStringSerializer.class) Long id,
        String name,
        ModelProtocol protocol,
        String baseUrl,
        String modelName,
        boolean isDefault,
        boolean isKnowledgeDefault) {

    public ModelConfig(Long id, String name, ModelProtocol protocol, String baseUrl, String modelName, boolean isDefault) {
        this(id, name, protocol, baseUrl, modelName, isDefault, false);
    }
}