# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SupportFlow AI API
  version: 1.0.0-demo
  description: |
    多租户电商售后支持平台。受保护接口从 JWT 推导 tenantId，不接受客户端传入的 tenantId。
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/v1/tenants/register:
    post:
      tags: [identity]
      security: []
      summary: 注册租户管理员（首次使用可不提供工作区代码）
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TenantRegistrationRequest'}}}}
      responses: {'201': {description: 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/RegistrationResponse'}}}}, '400': {$ref: '#/components/responses/Problem'}}
  /api/v1/customers/register:
    post:
      tags: [identity]
      security: []
      summary: 注册消费者
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CustomerRegistrationRequest'}}}}
      responses: {'201': {description: 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/RegistrationResponse'}}}}, '400': {$ref: '#/components/responses/Problem'}}
  /api/v1/auth/login:
    post:
      tags: [identity]
      security: []
      summary: 登录并获得令牌
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/LoginRequest'}}}}
      responses: {'200': {description: 令牌对, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}, '401': {$ref: '#/components/responses/Problem'}}
  /api/v1/auth/refresh:
    post:
      tags: [identity]
      security: []
      summary: 轮换刷新令牌
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RefreshTokenRequest'}}}}
      responses: {'200': {description: 新令牌对, content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}}}
  /api/v1/auth/session:
    get:
      tags: [identity]
      summary: 当前会话身份
      responses: {'200': {description: 身份, content: {application/json: {schema: {$ref: '#/components/schemas/Session'}}}}, '401': {$ref: '#/components/responses/Problem'}}
  /api/v1/auth/profile:
    get:
      tags: [identity]
      summary: 查询当前账户和工作区资料
      responses: {'200': {description: 当前账户资料, content: {application/json: {schema: {$ref: '#/components/schemas/CurrentUserProfile'}}}}, '401': {$ref: '#/components/responses/Problem'}}
    patch:
      tags: [identity]
      summary: 更新当前账户的显示名称
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateCurrentUserProfileRequest'}}}}
      responses: {'200': {description: 已更新, content: {application/json: {schema: {$ref: '#/components/schemas/CurrentUserProfile'}}}}, '400': {$ref: '#/components/responses/Problem'}}
  /api/v1/admin/members:
    get:
      tags: [identity]
      summary: 查询当前租户可分配的管理员、主管与坐席
      responses: {'200': {description: 可分配成员目录, content: {application/json: {schema: {type: array, items: {type: object, properties: {membershipId: {type: string}, displayName: {type: string}, role: {type: string}}}}}}}}
  /api/v1/customer/orders:
    get:
      tags: [customer]
      summary: 当前消费者订单
      responses: {'200': {description: 订单列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/CustomerOrder'}}}}}}
  /api/v1/customer/conversations:
    post:
      tags: [conversation]
      summary: 创建会话
      responses: {'201': {description: 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/Conversation'}}}}}
  /api/v1/customer/conversations/{conversationId}/messages:
    post:
      tags: [conversation]
      summary: 提交消费者消息并创建生成任务
      parameters: [{$ref: '#/components/parameters/ConversationId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/SubmitMessageRequest'}}}}
      responses: {'202': {description: 已受理, content: {application/json: {schema: {$ref: '#/components/schemas/Generation'}}}}, '409': {$ref: '#/components/responses/Problem'}}
  /api/v1/customer/generations/{generationId}/events:
    get:
      tags: [conversation]
      summary: 读取并重放生成 SSE 事件
      parameters: [{$ref: '#/components/parameters/GenerationId'}, {name: Last-Event-ID, in: header, schema: {type: string}}]
      responses: {'200': {description: SSE 事件流, content: {text/event-stream: {schema: {type: string}, example: "id: 1730000000000-0\nevent: text.delta\ndata: {\"text\":\"您好\"}\n\n"}}}}
  /api/v1/admin/tickets:
    get:
      tags: [tickets]
      summary: 当前租户工单队列
      responses: {'200': {description: 工单列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Ticket'}}}}}}
  /api/v1/admin/tickets/{ticketId}/claim:
    post:
      tags: [tickets]
      summary: 认领未分配工单
      parameters: [{$ref: '#/components/parameters/TicketId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {description: 已认领, content: {application/json: {schema: {$ref: '#/components/schemas/Ticket'}}}}, '409': {$ref: '#/components/responses/Problem'}}
  /api/v1/admin/tickets/{ticketId}/context:
    get:
      tags: [tickets]
      summary: 查询工单关联会话消息、知识与工具轨迹及消费者订单
      parameters: [{$ref: '#/components/parameters/TicketId'}]
      responses: {'200': {description: 坐席处理上下文}, '400': {$ref: '#/components/responses/Problem'}}
  /api/v1/admin/tickets/{ticketId}/assign:
    post:
      tags: [tickets]
      summary: 转派工单并记录分配审计
      parameters: [{$ref: '#/components/parameters/TicketId'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [membershipId], properties: {membershipId: {type: string, pattern: '^[0-9]+$'}}}}}}
      responses: {'200': {description: 已转派, content: {application/json: {schema: {$ref: '#/components/schemas/Ticket'}}}}, '400': {$ref: '#/components/responses/Problem'}, '409': {$ref: '#/components/responses/Problem'}}
  /api/v1/admin/tickets/{ticketId}/status:
    post:
      tags: [tickets]
      summary: 按状态机变更工单状态
      parameters: [{$ref: '#/components/parameters/TicketId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChangeTicketStatusRequest'}}}}
      responses: {'200': {description: 已变更, content: {application/json: {schema: {$ref: '#/components/schemas/Ticket'}}}}, '409': {$ref: '#/components/responses/Problem'}}
  /api/v1/admin/tickets/{ticketId}/comments:
    get:
      tags: [tickets]
      summary: 查询内部备注
      parameters: [{$ref: '#/components/parameters/TicketId'}]
      responses: {'200': {description: 备注列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/TicketComment'}}}}}}
    post:
      tags: [tickets]
      summary: 添加内部备注
      parameters: [{$ref: '#/components/parameters/TicketId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/TicketCommentRequest'}}}}
      responses: {'201': {description: 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/TicketComment'}}}}}
  /api/v1/admin/approvals:
    get:
      tags: [approvals]
      summary: 查询审批申请
      responses: {'200': {description: 审批列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Approval'}}}}}}
    post:
      tags: [approvals]
      summary: 创建包含标准化业务证据的审批申请
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateApprovalRequest'}}}}
      responses: {'201': {description: 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/Approval'}}}}, '400': {$ref: '#/components/responses/Problem'}}
  /api/v1/admin/approvals/{approvalId}/decision:
    post:
      tags: [approvals]
      summary: 审批高风险动作
      parameters: [{name: approvalId, in: path, required: true, schema: {type: string}}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, required: [decision], properties: {decision: {type: string, enum: [APPROVED, REJECTED]}}}}}}
      responses: {'200': {description: 审批结果, content: {application/json: {schema: {$ref: '#/components/schemas/Approval'}}}}, '409': {$ref: '#/components/responses/Problem'}}
  /api/v1/admin/approvals/{approvalId}/revoke:
    post:
      tags: [approvals]
      summary: 撤销仍处于待处理状态的审批申请
      parameters: [{name: approvalId, in: path, required: true, schema: {type: string}}, {$ref: '#/components/parameters/IdempotencyKey'}]
      responses: {'200': {description: 已撤销, content: {application/json: {schema: {$ref: '#/components/schemas/Approval'}}}}, '409': {$ref: '#/components/responses/Problem'}}
  /api/v1/admin/knowledge-bases:
    get:
      tags: [knowledge]
      summary: 查询当前租户知识库
      responses: {'200': {description: 知识库列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/KnowledgeBase'}}}}}}
    post:
      tags: [knowledge]
      summary: 创建知识库
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateKnowledgeBaseRequest'}}}}
      responses: {'201': {description: 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/KnowledgeBase'}}}}}
  /api/v1/admin/knowledge-bases/{knowledgeBaseId}/documents:
    get:
      tags: [knowledge]
      summary: 查询知识库文档
      parameters: [{$ref: '#/components/parameters/KnowledgeBaseId'}]
      responses: {'200': {description: 文档列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/KnowledgeDocument'}}}}}}
  /api/v1/admin/knowledge-bases/{knowledgeBaseId}/documents/upload:
    post:
      tags: [knowledge]
      summary: 上传并摄取知识文档
      parameters: [{$ref: '#/components/parameters/KnowledgeBaseId'}]
      requestBody: {required: true, content: {multipart/form-data: {schema: {type: object, required: [file], properties: {file: {type: string, format: binary}}}}}}
      responses: {'201': {description: 已上传, content: {application/json: {schema: {$ref: '#/components/schemas/KnowledgeDocument'}}}}, '409': {$ref: '#/components/responses/Problem'}}
  /api/v1/admin/knowledge-bases/{knowledgeBaseId}/documents/{documentId}/retry:
    post:
      tags: [knowledge]
      summary: 重试失败的知识文档摄取
      parameters: [{$ref: '#/components/parameters/KnowledgeBaseId'}, {name: documentId, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: 重试结果, content: {application/json: {schema: {$ref: '#/components/schemas/KnowledgeDocument'}}}}}
  /api/v1/admin/knowledge-bases/{knowledgeBaseId}/documents/rebuild:
    post:
      tags: [knowledge]
      summary: 重建当前租户知识库索引
      parameters: [{$ref: '#/components/parameters/KnowledgeBaseId'}]
      responses: {'200': {description: 重建结果}}
  /api/v1/admin/models:
    get:
      tags: [models]
      summary: 查询当前租户模型配置
      responses: {'200': {description: 模型配置列表, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ModelConfig'}}}}}}
    post:
      tags: [models]
      summary: 创建模型配置
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateModelConfigRequest'}}}}
      responses: {'201': {description: 已创建, content: {application/json: {schema: {$ref: '#/components/schemas/ModelConfig'}}}}}
  /api/v1/admin/models/probe:
    post:
      tags: [models]
      summary: 探测模型服务连通性
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ModelProbeRequest'}}}}
      responses: {'200': {description: 探测结果}}
  /api/v1/admin/models/{modelConfigId}/default:
    patch:
      tags: [models]
      summary: 切换当前租户默认聊天模型
      parameters: [{name: modelConfigId, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: 已切换，后续请求即时读取新配置, content: {application/json: {schema: {$ref: '#/components/schemas/ModelConfig'}}}}, '400': {$ref: '#/components/responses/Problem'}}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    ConversationId: {name: conversationId, in: path, required: true, schema: {type: string}}
    GenerationId: {name: generationId, in: path, required: true, schema: {type: string}}
    TicketId: {name: ticketId, in: path, required: true, schema: {type: string}}
    KnowledgeBaseId: {name: knowledgeBaseId, in: path, required: true, schema: {type: string}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1}}
  responses:
    Problem: {description: RFC 9457 problem response, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  schemas:
    TenantRegistrationRequest: {type: object, required: [email, displayName, password], properties: {tenantCode: {type: string, description: 可选；缺省时由服务端生成}, tenantName: {type: string, description: 可选；缺省时由显示名称生成}, email: {type: string, format: email}, displayName: {type: string}, password: {type: string, minLength: 12}}}
    CustomerRegistrationRequest: {type: object, required: [tenantCode, email, displayName, password], properties: {tenantCode: {type: string}, email: {type: string, format: email}, displayName: {type: string}, password: {type: string, minLength: 12}}}
    RegistrationResponse: {type: object, properties: {tenantId: {type: string}, userId: {type: string}, membershipId: {type: string}, tenantCode: {type: string, description: 管理员首次注册时由服务端生成的内部工作区标识}}}
    LoginRequest: {type: object, required: [tenantCode, email, password], properties: {tenantCode: {type: string}, email: {type: string, format: email}, password: {type: string}}}
    RefreshTokenRequest: {type: object, required: [refreshToken], properties: {refreshToken: {type: string}}}
    TokenPair: {type: object, required: [accessToken, refreshToken], properties: {accessToken: {type: string}, refreshToken: {type: string}}}
    Session: {type: object, properties: {userId: {type: string}, tenantId: {type: string}, membershipId: {type: string}, role: {type: string}}}
    CurrentUserProfile: {type: object, properties: {userId: {type: string}, tenantId: {type: string}, membershipId: {type: string}, displayName: {type: string}, email: {type: string, format: email}, role: {type: string}, tenantName: {type: string}, tenantCode: {type: string}, joinedAt: {type: string, format: date-time}}}
    UpdateCurrentUserProfileRequest: {type: object, required: [displayName], properties: {displayName: {type: string, minLength: 1, maxLength: 128}}}
    CustomerOrder: {type: object, properties: {orderNo: {type: string}, status: {type: string}, totalAmount: {type: number}, currency: {type: string}, createdAt: {type: string, format: date-time}}}
    Conversation: {type: object, properties: {id: {type: string}, status: {type: string}, createdAt: {type: string, format: date-time}}}
    SubmitMessageRequest: {type: object, required: [content], properties: {content: {type: string, minLength: 1}}}
    Generation: {type: object, properties: {id: {type: string}, conversationId: {type: string}, status: {type: string}, createdAt: {type: string, format: date-time}}}
    Ticket: {type: object, properties: {id: {type: string}, customerId: {type: string}, title: {type: string}, status: {type: string, enum: [NEW, OPEN, PENDING_CUSTOMER, PENDING_APPROVAL, RESOLVED, CLOSED]}, priority: {type: string, enum: [LOW, NORMAL, HIGH, URGENT]}, assignedMembershipId: {type: string}, resolutionDueAt: {type: string, format: date-time}}}
    ChangeTicketStatusRequest: {type: object, required: [status], properties: {status: {type: string, enum: [OPEN, PENDING_CUSTOMER, PENDING_APPROVAL, RESOLVED, CLOSED]}}}
    TicketCommentRequest: {type: object, required: [content], properties: {content: {type: string, minLength: 1}}}
    TicketComment: {type: object, properties: {id: {type: string}, ticketId: {type: string}, content: {type: string}, createdAt: {type: string, format: date-time}}}
    CreateApprovalRequest: {type: object, required: [actionType, actionSummary, orderNo, amount, currency, eligibilityEvidence], properties: {actionType: {type: string}, actionSummary: {type: string}, orderNo: {type: string}, amount: {type: number, minimum: 0.01}, currency: {type: string, pattern: '^[A-Za-z]{3}$'}, eligibilityEvidence: {type: string}}}
    Approval: {type: object, properties: {id: {type: string}, status: {type: string, enum: [PENDING, APPROVED, REJECTED, REVOKED, EXPIRED, EXECUTING, EXECUTED, FAILED]}, actionType: {type: string}, actionSummary: {type: string}, orderNo: {type: string}, amount: {type: number}, currency: {type: string}, eligibilityEvidence: {type: string}, requestedByMembershipId: {type: string}, decidedByMembershipId: {type: string}, version: {type: integer, format: int64}, expiresAt: {type: string, format: date-time}}}
    CreateKnowledgeBaseRequest: {type: object, required: [name], properties: {name: {type: string}, description: {type: string}}}
    KnowledgeBase: {type: object, properties: {id: {type: string}, name: {type: string}, description: {type: string}, status: {type: string}, version: {type: integer, format: int64, minimum: 1}}}
    KnowledgeDocument: {type: object, properties: {id: {type: string}, fileName: {type: string}, contentHash: {type: string}, status: {type: string}}}
    CreateModelConfigRequest: {type: object, required: [name, protocol, baseUrl, modelName, apiKey], properties: {name: {type: string}, protocol: {type: string, enum: [OPENAI_COMPATIBLE, ANTHROPIC_MESSAGES]}, baseUrl: {type: string, format: uri}, modelName: {type: string}, apiKey: {type: string, writeOnly: true}, isDefault: {type: boolean}}}
    ModelProbeRequest: {type: object, required: [baseUrl, apiKey], properties: {baseUrl: {type: string, format: uri}, apiKey: {type: string, writeOnly: true}}}
    ModelConfig: {type: object, properties: {id: {type: string}, name: {type: string}, protocol: {type: string}, baseUrl: {type: string}, modelName: {type: string}, isDefault: {type: boolean}}}
    Problem: {type: object, properties: {type: {type: string, format: uri}, title: {type: string}, status: {type: integer}, detail: {type: string}, code: {type: string}, requestId: {type: string}}}